requirements.txt
# ===================================================================
# Health & Wellness Predictor - Backend Dependencies
# Minimal requirements for production deployment (no training needed)
# ===================================================================

# Core Data Science (for ML model loading)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0

# Machine Learning - Gradient Boosting Support
# Note: Only XGBoost and LightGBM (no CatBoost to avoid cmake dependency)
xgboost>=2.0.0
lightgbm>=4.0.0

# API Framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic[email]>=2.4.0,<3.0.0

# Authentication & Security
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6

# Environment Configuration
python-dotenv>=1.0.0
notebook>=7.0.0

# Utility Libraries
python-dotenv>=1.0.0
requests>=2.31.0

# Progress Bars
tqdm>=4.66.0

# Date/Time Utilities
python-dateutil>=2.8.0

# Email validation
email-validator>=2.1.0

